+++ src/simtellus/start-simtellus.sh
@@ -1,14 +1,14 @@
 #!/bin/sh
 
-# Install necessary gems
+# install deps
 apk add --no-cache build-base curl
 
-# Function to check if a gem is installed
+# is a gem is installed?
 gem_installed() {
   gem list -i "$1" > /dev/null 2>&1
 }
 
-# Install necessary gems if not already installed
+# Install gem deps if not installed
 if ! gem_installed sinatra || ! gem_installed puma; then
   gem install sinatra json dotenv rack numeric rackup puma
 else
@@ -23,26 +23,34 @@ set +a
 # Create the log directory if it doesn't exist, using LOG_PATH
 mkdir -p "$(dirname "$LOG_PATH")"
 
-# Run the simulation server in the foreground
+# Run server in the background
 ruby /app/server.rb &
 SIMTELLUS_PID=$!
 
-# Wait for the server to start (adjust the sleep duration as needed)
-sleep 28
-
 # Define the URL for the curl request
 URL="http://localhost:${SIMTELLUS_PORT}/planet_state"
-
-# Check the endpoint once
-response=$(curl -s -w "%{http_code}" -o /dev/null "$URL")
-str="[ $0 server self-test ]"
-if [ "$response" -eq 200 ]; then
-  echo "$str 😎 $response."
+str="[ $0 test server ]"
+# Check if API_KEY_AUTH_ENABLE is set to 'true'
+if [ "$API_KEY_AUTH_ENABLE" = "true" ]; then
+  echo "$str API_KEY_AUTH_ENABLE true, self-test skipped."
 else
-  echo "$str FAIL: Endpoint not responding."
-  kill $SIMTELLUS_PID
-  exit 1
+  # Wait for the server to start
+  sleep 28
+
+  # Perform the curl request to test the endpoint
+  response=$(curl -s -w "%{http_code}" -o /dev/null "$URL")
+
+  if [ "$response" -eq 200 ]; then
+    echo "$str 😎 $response."
+  else
+    echo "$str FAIL: Endpoint not responding."
+    kill $SIMTELLUS_PID
+    exit 1
+  fi
 fi
+
+# end upon signal
+trap 'kill $SIMTELLUS_PID' TERM INT
 
 # keep the planet running
 wait $SIMTELLUS_PID