+++ src/webserver/deps.ts
@@ -1,0 +1,2 @@
+export { Application, Router } from "https://deno.land/x/oak@v12.5.0/mod.ts";
+export { Client } from "https://deno.land/x/postgres@v0.17.0/mod.ts";